AppCertificate/Portfolio/requirements.txt
flask
mysql-connector-python
datetime
flask_session
